Add files via upload
Revolutionizing Automation: The Journey of Our Robotic Arm Project

I am thrilled to share an exciting project that has been a cornerstone of our innovation journey: the development of a robotic arm designed to enhance automation .

From this project we were able to learn and use Ardunio, LINUX, ROS, Motion Planning, CAD, MATLAB, Gazebo , Moveit Setup Assistant .With the help of this we were able to create a highly accurate robotic arm with inverse kinematics, enhancing operational efficiency by 48%.

Equipped with high-precision servo motors and sensors, our robotic arm offers smooth and accurate movement across multiple axes. This ensures precise manipulation of objects .As we look to the future, our focus is on continuous improvement and innovation. We are exploring advanced AI capabilities, further enhancing the robotic arm’s adaptability and intelligence. Additionally, we are working on expanding its application scope, making it a versatile tool and make it useful for a normal person in daily life .